+++ packages/atlas/src/joystream-lib/utils.ts
@@ -68,6 +68,13 @@ export const calculateAssetsBloatFee = (
 }
 
 export const parseAccountBalance = (balances: DeriveBalancesAll): AccountBalanceInfo => {
+  /*
+    balances.freeBalance = all the tokens in the account
+    feeUsable = balance usable for paying fees
+    more here: 
+    https://gist.github.com/Lezek123/88b85b6af866feaa4f6b5064ce528a93
+  */
+
   const feeUsable = balances.freeBalance.sub(balances.frozenFee)
 
   const lockedBalance = feeUsable.sub(balances.availableBalance)