+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed issue with displaying incorrect locked funds
+
 ## [1.2.1] - 2022-12-27
 
 ### Fixed