+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+## [1.2.2] - 2023-01-11
+
+### Added
+
 - Welcome dialog when content focus is present
 - Fallback state for video hero and video category hero